Move podcast logic into stitch and use at build time
