+++ src/utils/devhub-api-stitch.js
@@ -9,6 +9,11 @@ const callDevhubAPIStitchFunction = async (fnName, ...fnArgs) => {
     }
 };
 
+export const requestLybsinPodcasts = async () => {
+    const result = await callDevhubAPIStitchFunction('fetchLybsinPodcasts');
+    return result;
+};
+
 export const requestMDBTwitchStream = async () => {
     const result = await callDevhubAPIStitchFunction('fetchMDBTwitchStream');
     return result;
