+++ src/utils/setup/fetch-build-time-media.js
@@ -1,18 +1,23 @@
 import { initStitch } from './init-stitch';
 import { STITCH_AUTH_APP_ID } from '../../constants';
+import { parsePodcasts } from '../parse-podcasts';
 import { simplifyTwitchResponse } from '../simplify-twitch-response';
 import { simplifyYoutubeResponse } from '../simplify-youtube-response';
 
 const MAX_RESULTS = 5;
 
-export const fetchBuildTimeVideos = async () => {
+export const fetchBuildTimeMedia = async () => {
     const client = await initStitch(STITCH_AUTH_APP_ID);
-    const [youtubeVideos, twitchVideos] = await Promise.all([
+    const [youtubeVideos, twitchVideos, lybsinPodcasts] = await Promise.all([
         client.callFunction('fetchYoutubeData', [MAX_RESULTS]),
         client.callFunction('fetchMDBTwitchVideos', [MAX_RESULTS]),
+        client.callFunction('fetchLybsinPodcasts', []),
     ]);
-    return [
-        youtubeVideos.items.map(simplifyYoutubeResponse),
-        twitchVideos.data.map(simplifyTwitchResponse),
-    ].flat();
+    return {
+        videos: [
+            youtubeVideos.items.map(simplifyYoutubeResponse),
+            twitchVideos.data.map(simplifyTwitchResponse),
+        ].flat(),
+        podcasts: parsePodcasts(lybsinPodcasts),
+    };
 };
